requirements.txt
Flask
Flask-Mail
Flask-Limiter
python-dotenv
gunicorn
requests
APScheduler
Pillow
flask-minify
htmlmin
